requirements.txt
# ====================================

# SILAHKAN MASUKKAN PACKAGE YANG DI GUNAKAN DISINI

# ====================================

autopep8==1.5
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.0
timeago==1.0.15
MarkupSafe==1.1.1
pycodestyle==2.5.0
Werkzeug==0.16.1
firebase-admin==4.5.2
Flask-HTTPAuth==4.1.0
Flask-Mail==0.9.0
requests==2.25.1
python-dotenv==0.15.0
openpyxl==3.0.5
midtransclient==1.2.0
Pyrebase4==4.4.3